requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
discord.py @ git+https://github.com/Rapptz/discord.py@55849d996e65613a334d73adbfc43bbe7b77d31a
Flask==2.0.1
html2image @ git+https://github.com/vgalin/html2image@54166da7b2df0e4a0bc94d335aef3363d17880cc
idna==3.2
importlib-metadata==4.12.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
oauthlib==3.1.1
pycparser==2.20
pyee==8.2.2
pyOpenSSL==20.0.1
python-dotenv==0.18.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
tqdm==4.64.0
typing-extensions==3.10.0.0
urllib3==1.26.6
websocket-client==1.2.3
websockets==10.3
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.8.0
